backend/pyproject.toml
[project]
name = "cobblemon-academy-tracker-api"
version = "0.1.0"
description = ""
authors = [
    {name = "axelfrache",email = "axelfrache@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "motor (>=3.7.1,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "cobblemon_academy_tracker_api", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
